UPD setup-openshift - added check if cluster was stoped

+++ systemtests/scripts/setup-openshift.sh
@@ -3,6 +3,12 @@ SCRIPTDIR=`dirname $0`
 
 oc cluster down #for the case that cluster is already running
 
+oc get status
+if [ $? -ne 0 ]; then
+    echo "ERROR: shutting down of openshift cluster failed, tests won't be executed"
+    exit 1
+fi
+
 sudo rm -rf /var/lib/origin/openshift.local.pv
 sudo rm -rf /var/log/containers/*
 sudo rm -rf /var/log/pods/*
